Added formula for mecab, a Japanese morphological analyser
mecab is the main application that does morphological analysis on Japanese text.
mecab-ipadic is data that is required by mecab to work. This is the recommended data file by the author, but other data files are available on the mecab homepage.

This formula sets utf-8 as the default charset for mecab-ipadic, but this can be changed with the --with-charset=[sjis|euc-jp] flag.